requirements.txt
fitz
pymupdf
nltk
# spacy
sentence-transformers
faiss-cpu
sumy
networkx
bertopic
llama-cpp-python
scikit-learn
hdbscan
langchain
langchain_community
markdown-it-py
